+++ outputs.tf
@@ -48,19 +48,3 @@ output "mysql_vnet_rules" {
   value       = azurerm_mysql_virtual_network_rule.vnet_rules
   description = "The map of all vnet rules"
 }
-
-output "mysql_databases_users" {
-  description = "Map of user name for each database"
-  value       = { for db, c in var.databases : db => mysql_user.users[db].user if length(mysql_user.users) > 0 }
-}
-
-output "mysql_databases_logins" {
-  description = "Map of user login for each database"
-  value       = { for db, c in var.databases : db => format("%s@%s", mysql_user.users[db].user, azurerm_mysql_server.mysql_server.name) if length(mysql_user.users) > 0 }
-}
-
-output "mysql_databases_passwords" {
-  description = "Map of user password for each database"
-  value       = { for db, c in var.databases : db => random_password.db_passwords[db].result if length(random_password.db_passwords) > 0 }
-  sensitive   = true
-}